[activation] Support in-place activation layer
This patch allows support of in-place activation layer.
With this patch, the input/output of activation layer are not allocated
as activation layer can work with its output and does not need input.
So, the input allocated for the next layer can be reused here and activation
layer can work in-place.

**Self evaluation:**
1. Build test: [x]Passed [ ]Failed [ ]Skipped
2. Run test: [x]Passed [ ]Failed [ ]Skipped

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>